+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- All credits go to Dane Mackier of FilledStacks.com for the original Stacked package.
+- GitHub: https://github.com/FilledStacks/stacked
+- Pub Dev: https://pub.dev/packages/stacked
+- This package contains only the basic view model, view model builder and view model widget with stripped down, added and altered functionality.